gtkgui: Increase entry box width and add alignment
Due to the dot in the float thickness (1.0), a width of 2 chars was too
small.

It is increased to 4 chars, to allow a nice display of double
digit widths (10.0).

Also the text inside the Gtk.Entry is flushed right, to have a nice
display of floats with less than 4 chars width.

Fixes: #43