Adding a set of methods and tracking variables for supporting the pinning of pictures in the tool.
Parent pages are maintained for each subsequent level and the child page has access to all it's parents

For the time being the child only retrieves the pinned pictures from the parents. We also paint a delete button on the UI based on whether the image is pinned or otherwise.